Create separate crate for converting parser CST into AST
