+++ Cargo.toml
@@ -45,5 +45,5 @@ members = [
   "galvan-transpiler",
   "galvan-test-macro",
   "galvan-test",
-  "tree-sitter-galvan",
+  "tree-sitter-galvan", "galvan-into-ast",
 ]
