+++ galvan-ast/src/result.rs
@@ -1,6 +1,4 @@
 use derive_more::From;
-use from_pest::{ConversionError, Void};
-use galvan_pest::ParserError;
 use thiserror::Error;
 
 use crate::Ast;
@@ -11,10 +9,10 @@ pub type AstResult = Result<Ast>;
 
 #[derive(Debug, Error, From)]
 pub enum AstError {
-    #[error("Error when converting parsed code to AST: {0}")]
-    ConversionError(ConversionError<Void>),
-    #[error("Error when parsing code: {0}")]
-    ParseError(ParserError),
+    #[error("Error when converting parsed code to AST")]
+    ConversionError,
+    #[error("Error when parsing code")]
+    ParseError,
     #[error("Duplicate main function")]
     DuplicateMain,
 }
